# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Alpine Ecommerce API",
    "version": "1.0.0",
    "description": "# Alpine Ecommerce Backend API\n\nComplete API documentation for Alpine Ecommerce e-commerce platform.\n\n## Features\n- 🔐 JWT Authentication\n- 📦 Product Management\n- 🛒 Order Processing\n- 💳 Payment Integration (Razorpay)\n- 👥 Customer Management\n- 📊 Dashboard & Reports\n- 📝 Error Logging\n\n## Authentication\nMost endpoints require JWT authentication. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your-token>\n```\n\n## Base URL\n- **Current Server**: http://13.235.53.15:8081/api/v1\n- **Configuration**: \n  - Set `API_BASE_URL` in .env for full control (recommended)\n  - Or set `PRODUCTION_URL` for production domain\n  - Or set `DOMAIN`, `PORT`, and `PROTOCOL` for auto-construction\n\n## Response Format\nAll responses follow a consistent format:\n```json\n{\n  \"responseStatusCode\": 200,\n  \"isError\": false,\n  \"errorData\": null,\n  \"successData\": { ... }\n}\n```\n\n## Error Handling\nErrors are returned with appropriate status codes:\n```json\n{\n  \"responseStatusCode\": 400,\n  \"isError\": true,\n  \"errorData\": {\n    \"displayMessage\": \"Error description\",\n    \"apiErrorType\": 1\n  }\n}\n```",
    "contact": {
      "name": "Alpine Ecommerce API Support",
      "email": "support@wildvalleyfoods.in"
    },
    "license": {
      "name": "Proprietary"
    },
    "servers": [
      {
...
      }
    ],
...
        }
...
  },
  "paths": {
...
    "/api/v1/banner/create": {
      "post": {
        "tags": [
          "Banners"
        ],
        "summary": "Create a new banner",
        "description": "Creates a new banner with a single image upload",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "reqData",
                  "imagePath"
                ],
                "properties": {
                  "reqData": {
                    "type": "string",
                    "format": "json",
                    "description": "JSON string containing banner data",
                    "example": "{\"title\":\"Summer Sale\",\"type\":\"homepage\",\"link\":\"/products\"}"
                  },
                  "imagePath": {
                    "type": "string",
                    "format": "binary",
                    "description": "Banner image file"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Banner created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/banner/getall/paginated": {
      "get": {
        "tags": [
          "Banners"
        ],
        "summary": "Get paginated banners",
        "description": "Returns banners with pagination support",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated banners retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/banner/getbyid/{id}": {
      "get": {
        "tags": [
          "Banners"
        ],
        "summary": "Get banner by ID",
        "description": "Returns a specific banner by its ID",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Banner ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Banner retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/banner/getbytype/{type}": {
      "get": {
        "tags": [
          "Banners"
        ],
        "summary": "Get banners by type",
        "description": "Returns banners filtered by type",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "type",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Banner type (e.g., homepage, category)"
          }
        ],
        "responses": {
          "200": {
            "description": "Banners retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/banner/update/{id}": {
      "put": {
        "tags": [
          "Banners"
        ],
        "summary": "Update banner",
        "description": "Updates an existing banner with optional image upload",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Banner ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "reqData": {
                    "type": "string",
                    "format": "json",
                    "description": "JSON string containing updated banner data",
                    "example": "{\"title\":\"Updated Banner\",\"type\":\"homepage\"}"
                  },
                  "imagePath": {
                    "type": "string",
                    "format": "binary",
                    "description": "New banner image file (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Banner updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/banner/delete/{id}": {
      "delete": {
        "tags": [
          "Banners"
        ],
        "summary": "Delete banner",
        "description": "Deletes a banner by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Banner ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Banner deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /register or /login endpoint. Enter your token (with or without \"Bearer \" prefix)"
      }
    }
  },
...
}